requirements.txt
# Core dependencies required for data preparation, training, explainability,
# and the Streamlit user interface.
torch
torchvision
numpy
pandas
matplotlib
tqdm
scikit-learn
opencv-python
streamlit
pillow
plotly
reportlab
